fix validators
